+++ src/elements/tab/index.ts
@@ -1,1 +1,1 @@
-export * from './tabs';
+export * from './Tabs';
