+++ src/elements/tab/tab.tsx
@@ -2,8 +2,8 @@ import { Tab as HeadlessTab } from '@headlessui/react';
 import clsx from 'clsx';
 import type { ComponentProps } from 'react';
 import { Fragment, useContext } from 'react';
-import { TabContext } from './tab-provider';
-import { VARIANTS } from './tabs';
+import { TabContext } from './TabProvider';
+import { VARIANTS } from './Tabs';
 
 type TabProps = ComponentProps<'button'> & { noText?: boolean };
